+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Version 0.4.4
 * Added full support for the LabOne Scope module (no need to fallback to zhinst.core)
 * Renamed `zhinst.toolkit.nodetree.nodes.WildcardResult` to `zhinst.toolkit.nodetree.helpers.NodeDict`
+* Added `active_validation` argument to `CommandTable`. By disabling it, `CommandTable` does not actively 
+  validate the inputs and therefore it improves the speed for command table creation.
 
 ## Version 0.4.3
 * Fix issue that prevented correct compilation of sequences for AWG cores other than the first one.